move change password to settings
