--- frontend/static/css/2FA.css
@@ -1,21 +1,0 @@
-#two-factor-section {
-  background-color: #1a1a1a;
-  border: 1px solid #00cc99;
-  padding: 20px;
-  border-radius: 10px;
-  box-shadow: 0px 0px 15px rgba(0, 204, 153, 0.3);
-  max-width: 600px;
-  margin: 20px auto;
-}
-
-#otp_code {
-  background-color: #1a1a1a;
-  color: #00cc99;
-}
-
-.qrcode-image {
-  width: 200px;
-  height: auto;
-  display: block;
-  margin: 0 auto;
-}
