+++ frontend/static/index.html
@@ -33,7 +33,7 @@
               <ul class="dropdown-menu dropdown-menu-end" aria-labelledby="navbarDropdown">
                 <li><a class="dropdown-item" href="/profile">Profile</a></li>
                 <li><a class="dropdown-item" href="/friends">Friends</a></li>
-                <li><a class="dropdown-item" href="/profile/2FA">2FA</a></li>
+                <li><a class="dropdown-item" href="/settings">Settings</a></li>
                 <li><hr class="dropdown-divider"></li>
                 <li><a class="dropdown-item" href="/logout">Logout</a></li>
               </ul>
