--- frontend/static/pages/2FA.html
@@ -1,25 +1,0 @@
-
-<div class="container-settings">
-  <h1 class="mb-4">Settings</h1>
-  <div id="alert-settings"></div>
-  <div id="two-factor-section">
-    <h2 class="mb-3">Security</h2>
-    <button id="enable-2fa-button" class="btn btn-secondary mt-3">Enable 2FA</button>
-    <div id="qrcode-container" class="mt-3 d-none">
-      <img id="qrcode-image" src="" alt="QR Code for 2FA" class="qrcode-image" />
-      <div id="two-factor-activation-container">
-        <form id="two-factor-activation-form">
-          <div class="mb-3 text-start">
-            <label for="otp_code" class="form-label">Enter the 2FA code generated by your authenticator app to confirm activation:</label>
-            <input type="text" class="form-control" id="otp_code" name="otp_code" placeholder="Enter 2FA code"/>
-          </div>
-          <div class="d-flex gap-2">
-            <button type="submit" class="btn btn-primary">Confirm</button>
-            <button type="button" id="cancel-2fa-activation-btn" class="btn btn-secondary">Cancel</button>
-          </div>
-        </form>
-        <div id="activation-error" class="alert alert-danger d-none mt-2"></div>
-      </div>
-    </div>
-  </div>
-</div>
