+++ frontend/static/pages/edit-profile.html
@@ -30,30 +30,4 @@
       </div>
     </div>
   </div>
-
-  <!-- Section de changement de mot de passe sous forme de formulaire -->
-  <div id="edit-password" class="card mb-4">
-    <div class="edit-content">
-      <div class="edit-info">
-        <h2 class="mt-4">Change Password</h2>
-        <form id="change-password-form">
-          <div id="error-alert" class="alert alert-danger d-none" role="alert"></div>
-          <div class="mb-3-edit-profile text-start">
-            <label for="current-password"><strong>Current Password:</strong></label>
-            <input type="password" id="current-password" class="edit-control" placeholder="Enter current password">
-          </div>
-          <div class="mb-3-edit-profile text-start">
-            <label for="new-password"><strong>New Password:</strong></label>
-            <input type="password" id="new-password" class="edit-control" placeholder="Enter new password">
-          </div>
-          <div class="mb-3-edit-profile text-start">
-            <label for="confirm-new-password"><strong>Confirm New Password:</strong></label>
-            <input type="password" id="confirm-new-password" class="edit-control" placeholder="Confirm new password">
-            <div class="invalid-feedback">Passwords do not match.</div>
-          </div>
-          <button type="submit" id="save-password" class="btn btn-primary">Change Password</button>
-        </form>
-      </div>
-    </div>
-  </div>
 </div>
