+++ frontend/static/pages/settings.html
@@ -1,0 +1,51 @@
+
+<div class="container container-settings">
+  <h1 class="mb-4">Settings</h1>
+  <div id="alert-settings"></div>
+  <div id="two-factor-section">
+    <h2 class="mb-3">Security</h2>
+    <button id="enable-2fa-button" class="btn btn-secondary mt-3">Enable 2FA</button>
+    <div id="qrcode-container" class="mt-3 d-none">
+      <img id="qrcode-image" src="" alt="QR Code for 2FA" class="qrcode-image img-fluid" />
+      <div id="two-factor-activation-container">
+        <form id="two-factor-activation-form">
+          <div class="mb-3 text-start">
+            <label for="otp_code" class="form-label">Enter the 2FA code generated by your authenticator app to confirm activation:</label>
+            <input type="text" class="form-control" id="otp_code" name="otp_code" placeholder="Enter 2FA code"/>
+          </div>
+          <div class="d-flex gap-2">
+            <button type="submit" class="btn btn-primary">Confirm</button>
+            <button type="button" id="cancel-2fa-activation-btn" class="btn btn-secondary">Cancel</button>
+          </div>
+        </form>
+        <div id="activation-error" class="alert alert-danger d-none mt-2"></div>
+      </div>
+    </div>
+  </div>
+  <!-- Section de changement de mot de passe sous forme de formulaire -->
+  <div id="edit-password" class="card mb-4">
+    <div class="edit-content">
+      <div class="edit-info">
+        <h2 class="mt-4">Change Password</h2>
+        <form id="change-password-form">
+          <div id="error-alert" class="alert alert-danger d-none" role="alert"></div>
+          <div class="mb-3-edit-profile text-start">
+            <label for="current-password"><strong>Current Password:</strong></label>
+            <input type="password" id="current-password" class="edit-control" placeholder="Enter current password">
+          </div>
+          <div class="mb-3-edit-profile text-start">
+            <label for="new-password"><strong>New Password:</strong></label>
+            <input type="password" id="new-password" class="edit-control" placeholder="Enter new password">
+          </div>
+          <div class="mb-3-edit-profile text-start">
+            <label for="confirm-new-password"><strong>Confirm New Password:</strong></label>
+            <input type="password" id="confirm-new-password" class="edit-control" placeholder="Confirm new password">
+            <div class="invalid-feedback">Passwords do not match.</div>
+          </div>
+          <button type="submit" id="save-password" class="btn btn-primary">Change Password</button>
+        </form>
+      </div>
+    </div>
+  </div>
+</div>
+
